requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
motor==3.3.1
pinecone-client==2.2.4
python-dotenv==1.0.0
pydantic==2.4.2
pymongo==4.6.0
python-linkedin-v2==0.9.4
PyPDF2==3.0.1
langchain==0.0.335
tiktoken==0.5.1
pydantic-settings==2.0.3
openai==1.3.5
Flask==2.3.3
Flask-JWT-Extended==4.5.2
Flask-Mail==0.9.1
Flask-CORS==4.0.0
Werkzeug==2.3.7
bcrypt==4.0.1
